Added redis auth, fixed tests
