+++ internal/models/DbConnection.go
@@ -5,5 +5,8 @@ type DbConnection struct {
 	SqliteFileName string
 	RedisUrl       string
 	RedisPrefix    string
+	RedisUsername  string
+	RedisPassword  string
+	RedisUseSsl    bool
 	Type           int
 }
